Solved a problem on frmMain where the form would not retrieve the lake details into the frmLakeTypeRef subform
The events form has a subform to pull lake type and description from
tblPonds into a subform frmLakeTypeRef.  After a pondname was selected
from a dropdown the subform was not pulling the lake description data.
I wrote some code in cboPondName_AfterUpdate to mark the record, refresh
the form and then move back to the record, thus refreshing the subform.
Clunky but it's the only way I could get it to work.